complication: Change SgvComplication to show BG delta
Previously, the short title of the SGV complication displayed the time since the last reading using `TimeDifferenceComplicationText`.

This commit changes the short title to display the BG delta alongside the time since the last reading. A new setting, `key_show_detailed_delta`, is used to determine whether to show the detailed or regular delta value. This provides more relevant information at a glance.

To support this, a new helper function `shortTimeWithDelta()` is added to `DisplayFormat.kt`.